refactor: RandomReader
